+++ internal/hash/hash.go
